requirements.txt
streamlit
ultralytics
transformers
ftfy
accelerate
safetensors
torch
torchvision
pillow
opencv-python
numpy
plotly
altair
tensorflow   # only if you want MobileNet classification locally
